backend/requirements.txt
langchain
langchain-community
faiss-cpu
sentence-transformers
python-dotenv
fastapi
uvicorn
pymongo